pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "clkan"
version = "0.0.0"

requires-python = ">=3.12"

dependencies = [
    "torchvision==0.19.0",
    "torcheval==0.0.7",
    "torchaudio==2.4.0",
    "torch==2.4.0",
    "SQLAlchemy==1.4.46",
    "sklego==0.0",
    "seaborn==0.13.2",
    "pymysql==1.1.1",
    "pydantic==2.8.2",
    "pandas==2.2.2",
    "optuna>=3.1.0",
    "optuna-dashboard==0.16.0",
    "lightning==2.4.0",
    # TODO: Because of a version conflict with optuna we need to use an unreleased
    # version of hydra-optuna-sweeper. This should be removed once the new version
    # is released. See: https://github.com/facebookresearch/hydra/pull/2360
    "hydra-optuna-sweeper @ git+https://github.com/facebookresearch/hydra@8f97f8b0c8344048c3b3cf8d9c36495ebf3ce463#subdirectory=plugins/hydra_optuna_sweeper",
    "hydra-joblib-launcher==1.2.0",
    "hydra-core==1.3.2",
    "h5py==3.11.0",
    "latextable==1.0.1",
    "jupyter==1.0.0",
    "wandb==0.17.7"
]

[project.scripts]
clkan = "clkan.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/clkan", "src/efficient_kan", "src/claiutil"]
artifacts = ["*.yaml"]